refactor(common): share the durability helpers the journal needs
`irlume-camera` has no atomic-write, fsync-chain or checksum helper it can
reach, and the IR emitter's undo journal needs all three with exactly the
guarantees the login transactions already established.

Move `sha256_hex`, `ancestor_chain`, `fsync_ancestors`, `fsync_dir` and
`restrict` to `irlume-common` and have `logintx` call them there. Behaviour is
unchanged; the test that pins the sync chain moved with the function it covers,
because that list is the only part of an fsync a test can observe.

`sha256_hex` already existed twice: once in `thirdparty` describing itself as
"the one place the checksum is computed", and once in `logintx`. That claim was
already false and a third copy would have made it worse, so `thirdparty` now
re-exports the crate-root one and its callers keep their path.

Adds `remove_durable`, the counterpart to `write_0600_atomic` for a record whose
whole meaning is "there is unfinished business here": an unlink still in the
page cache when the machine loses power brings the record back, and a record
that comes back is acted on again.
