+++ crates/irlume-cli/src/logintx.rs
@@ -196,14 +196,7 @@ fn schema_version_1() -> u32 {
     1
 }
 
-/// Hex sha256 of a byte slice.
-pub(crate) fn sha256_hex(bytes: &[u8]) -> String {
-    use sha2::{Digest as _, Sha256};
-    Sha256::digest(bytes)
-        .iter()
-        .map(|b| format!("{b:02x}"))
-        .collect()
-}
+pub(crate) use irlume_common::sha256_hex;
 
 /// The digest of a file's current content, or `None` when it does not exist.
 ///
@@ -551,63 +544,12 @@ pub(crate) fn is_valid_id(id: &str) -> bool {
     id.len() == 32 && id.chars().all(|c| c.is_ascii_hexdigit())
 }
 
-/// Make every directory above `dir` durable, so the names leading to it survive
-/// a power loss.
-///
-/// Shallowest first, because a directory's entry lives in its parent: syncing
-/// `/var/lib/irlume` makes `login-transactions` findable, and does nothing for
-/// `irlume` itself, whose entry is in `/var/lib`. A record fsynced into a
-/// directory whose name did not survive is not a record.
-fn fsync_ancestors(dir: &Path) -> Result<(), String> {
-    for parent in ancestor_chain(dir) {
-        fsync_dir(&parent)?;
-    }
-    Ok(())
-}
-
-/// The directories to sync above `dir`, shallowest first.
-///
-/// Separated out because the interesting case cannot be observed from outside:
-/// whether an `fsync` happened is not visible in the filesystem afterwards, so
-/// the list is what a test can actually assert on.
-fn ancestor_chain(dir: &Path) -> Vec<PathBuf> {
-    let mut chain: Vec<PathBuf> = dir
-        .ancestors()
-        .skip(1) // `dir` itself is synced by the atomic write that fills it
-        .map(|p| {
-            // A relative path's last ancestor is "", which opens nothing. The
-            // directory a relative path is anchored in is the working
-            // directory, and that is where the entry actually lives. Filtering
-            // the empty one out instead left `IRLUME_STATE_DIR=state` syncing
-            // `state` while nothing synced the `state` entry itself.
-            if p.as_os_str().is_empty() {
-                PathBuf::from(".")
-            } else {
-                p.to_path_buf()
-            }
-        })
-        .collect();
-    chain.reverse();
-    chain
-}
-
-/// Make a directory's own contents durable, so entries created in it survive a
-/// power loss.
-///
-/// `fsync(2)` is explicit that syncing a file does not necessarily persist the
-/// directory entry naming it; the directory has to be synced too. Opening a
-/// directory read-only and syncing that descriptor is the way to do it.
-fn fsync_dir(dir: &Path) -> Result<(), String> {
-    std::fs::File::open(dir)
-        .and_then(|d| d.sync_all())
-        .map_err(|e| format!("fsync {}: {e}", dir.display()))
-}
-
-fn restrict(path: &Path, mode: u32) -> Result<(), String> {
-    use std::os::unix::fs::PermissionsExt;
-    std::fs::set_permissions(path, std::fs::Permissions::from_mode(mode))
-        .map_err(|e| format!("chmod {}: {e}", path.display()))
-}
+// The durability helpers this module was written around now live in
+// `irlume_common`, because `irlume-camera` needs the same guarantees for the IR
+// emitter's undo journal and two copies of an fsync chain drift apart in
+// exactly the way that makes one of them wrong. Behaviour is unchanged; the
+// tests that pin the sync chain moved with them.
+use irlume_common::{fsync_ancestors, fsync_dir, restrict};
 
 /// Why a surface could not be rolled back.
 #[derive(Debug, PartialEq, Eq)]
@@ -1052,49 +994,6 @@ mod tests {
             "1".repeat(32)
         ));
         assert_eq!(Transaction::load(&id).map(|r| r.schema_version), Ok(1));
-    }
-
-    /// Every directory whose entry has to survive is in the chain, shallowest
-    /// first, including the one a RELATIVE state root is anchored in.
-    ///
-    /// A directory's name lives in its parent, so syncing `state` does nothing
-    /// for the `state` entry itself; for a relative path that entry is in the
-    /// working directory. An earlier version dropped the empty last ancestor
-    /// instead of reading it as ".", which left exactly that gap. Asserted on
-    /// the list because an `fsync` leaves no trace in the filesystem to check.
-    #[test]
-    fn the_sync_chain_covers_every_directory_whose_entry_must_survive() {
-        let abs = ancestor_chain(Path::new("/var/lib/irlume/login-transactions"));
-        assert_eq!(
-            abs,
-            vec![
-                PathBuf::from("/"),
-                PathBuf::from("/var"),
-                PathBuf::from("/var/lib"),
-                PathBuf::from("/var/lib/irlume"),
-            ],
-            "shallowest first, and the store itself is left to the atomic write"
-        );
-
-        // The case the filter used to lose: nothing anchored `state`.
-        let rel = ancestor_chain(Path::new("state/login-transactions"));
-        assert_eq!(rel, vec![PathBuf::from("."), PathBuf::from("state")]);
-
-        // A store directly under a relative root still names the anchor.
-        assert_eq!(
-            ancestor_chain(Path::new("login-transactions")),
-            vec![PathBuf::from(".")]
-        );
-        // Nothing in a chain may be empty: an empty path opens nothing, so a
-        // sync of it is a sync that silently did not happen.
-        for dir in ["/a/b", "a/b", "b", "/"] {
-            assert!(
-                ancestor_chain(Path::new(dir))
-                    .iter()
-                    .all(|p| !p.as_os_str().is_empty()),
-                "{dir} produced an empty entry"
-            );
-        }
     }
 
     #[test]
