+++ crates/irlume-common/src/lib.rs
@@ -92,6 +92,91 @@ pub fn state_dir() -> std::path::PathBuf {
         .unwrap_or_else(|| std::path::PathBuf::from(STATE_DIR))
 }
 
+/// Hex sha256 of a byte slice.
+pub fn sha256_hex(bytes: &[u8]) -> String {
+    use sha2::{Digest as _, Sha256};
+    Sha256::digest(bytes)
+        .iter()
+        .map(|b| format!("{b:02x}"))
+        .collect()
+}
+
+/// Make every directory above `dir` durable, so the names leading to it survive
+/// a power loss.
+///
+/// Shallowest first, because a directory's entry lives in its parent: syncing
+/// `/var/lib/irlume` makes `login-transactions` findable, and does nothing for
+/// `irlume` itself, whose entry is in `/var/lib`. A record fsynced into a
+/// directory whose name did not survive is not a record.
+pub fn fsync_ancestors(dir: &std::path::Path) -> std::result::Result<(), String> {
+    for parent in ancestor_chain(dir) {
+        fsync_dir(&parent)?;
+    }
+    Ok(())
+}
+
+/// The directories to sync above `dir`, shallowest first.
+///
+/// Separated out because the interesting case cannot be observed from outside:
+/// whether an `fsync` happened is not visible in the filesystem afterwards, so
+/// the list is what a test can actually assert on.
+pub fn ancestor_chain(dir: &std::path::Path) -> Vec<std::path::PathBuf> {
+    let mut chain: Vec<std::path::PathBuf> = dir
+        .ancestors()
+        .skip(1) // `dir` itself is synced by the atomic write that fills it
+        .map(|p| {
+            // A relative path's last ancestor is "", which opens nothing. The
+            // directory a relative path is anchored in is the working
+            // directory, and that is where the entry actually lives. Filtering
+            // the empty one out instead left `IRLUME_STATE_DIR=state` syncing
+            // `state` while nothing synced the `state` entry itself.
+            if p.as_os_str().is_empty() {
+                std::path::PathBuf::from(".")
+            } else {
+                p.to_path_buf()
+            }
+        })
+        .collect();
+    chain.reverse();
+    chain
+}
+
+/// Make a directory's own contents durable, so entries created in it survive a
+/// power loss.
+///
+/// `fsync(2)` is explicit that syncing a file does not necessarily persist the
+/// directory entry naming it; the directory has to be synced too. Opening a
+/// directory read-only and syncing that descriptor is the way to do it.
+pub fn fsync_dir(dir: &std::path::Path) -> std::result::Result<(), String> {
+    std::fs::File::open(dir)
+        .and_then(|d| d.sync_all())
+        .map_err(|e| format!("fsync {}: {e}", dir.display()))
+}
+
+/// Set `path`'s permission bits, naming the path when it fails.
+pub fn restrict(path: &std::path::Path, mode: u32) -> std::result::Result<(), String> {
+    use std::os::unix::fs::PermissionsExt;
+    std::fs::set_permissions(path, std::fs::Permissions::from_mode(mode))
+        .map_err(|e| format!("chmod {}: {e}", path.display()))
+}
+
+/// Remove `path` and make its absence durable.
+///
+/// The counterpart to [`write_0600_atomic`] for a record whose whole meaning is
+/// "there is unfinished business here": an unlink still sitting in the page
+/// cache when the machine loses power brings the record back, and a record that
+/// comes back is acted on again. Already-gone is success, because the caller's
+/// postcondition is that nothing is there.
+pub fn remove_durable(path: &std::path::Path) -> std::result::Result<(), String> {
+    match std::fs::remove_file(path) {
+        Ok(()) => {}
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(()),
+        Err(e) => return Err(format!("remove {}: {e}", path.display())),
+    }
+    let dir = path.parent().unwrap_or_else(|| std::path::Path::new("."));
+    fsync_dir(dir)
+}
+
 /// Create or truncate `path` with mode 0600 and write `bytes`, then fsync.
 ///
 /// Mode-on-open (not write-then-chmod) so a secret-bearing file is never
@@ -640,6 +725,66 @@ pub(crate) mod testenv {
 
 #[cfg(test)]
 mod tests {
+    use std::path::{Path, PathBuf};
+
+    /// Every directory whose entry has to survive is in the chain, shallowest
+    /// first, including the one a RELATIVE state root is anchored in.
+    ///
+    /// A directory's name lives in its parent, so syncing `state` does nothing
+    /// for the `state` entry itself; for a relative path that entry is in the
+    /// working directory. An earlier version dropped the empty last ancestor
+    /// instead of reading it as ".", which left exactly that gap. Asserted on
+    /// the list because an `fsync` leaves no trace in the filesystem to check.
+    #[test]
+    fn the_sync_chain_covers_every_directory_whose_entry_must_survive() {
+        let abs = super::ancestor_chain(Path::new("/var/lib/irlume/login-transactions"));
+        assert_eq!(
+            abs,
+            vec![
+                PathBuf::from("/"),
+                PathBuf::from("/var"),
+                PathBuf::from("/var/lib"),
+                PathBuf::from("/var/lib/irlume"),
+            ],
+            "shallowest first, and the store itself is left to the atomic write"
+        );
+
+        // The case the filter used to lose: nothing anchored `state`.
+        let rel = super::ancestor_chain(Path::new("state/login-transactions"));
+        assert_eq!(rel, vec![PathBuf::from("."), PathBuf::from("state")]);
+
+        // A store directly under a relative root still names the anchor.
+        assert_eq!(
+            super::ancestor_chain(Path::new("login-transactions")),
+            vec![PathBuf::from(".")]
+        );
+        // Nothing in a chain may be empty: an empty path opens nothing, so a
+        // sync of it is a sync that silently did not happen.
+        for dir in ["/a/b", "a/b", "b", "/"] {
+            assert!(
+                super::ancestor_chain(Path::new(dir))
+                    .iter()
+                    .all(|p| !p.as_os_str().is_empty()),
+                "{dir} produced an empty entry"
+            );
+        }
+    }
+
+    /// A removal that is not durable brings the record back after a power loss,
+    /// and a record that comes back is acted on again. Absence is the whole
+    /// meaning of a resolved journal, so it gets the same treatment as a write.
+    #[test]
+    fn removing_a_record_that_is_already_gone_is_success() {
+        let dir = std::env::temp_dir().join("irlume-remove-durable-test");
+        std::fs::create_dir_all(&dir).unwrap();
+        let path = dir.join("record");
+        std::fs::write(&path, b"x").unwrap();
+        assert_eq!(super::remove_durable(&path), Ok(()));
+        assert!(!path.exists());
+        // Idempotent: a caller resuming after a crash must not fail here.
+        assert_eq!(super::remove_durable(&path), Ok(()));
+        let _ = std::fs::remove_dir_all(&dir);
+    }
 
     // --- cross-version wire compatibility for typed operation errors --------
     //
